Split 'tests/test_rio_features.py' into one file per command and rename to 'tests/test_rio_shapes.py'.
Missed a few things in the 'tests/test_rio_features.py' split.
